OSIS-4502 refactoring moved function fetch group elements for tree
	modified:   base/models/group_element_year.py
	modified:   base/tests/models/test_group_element_year.py
	modified:   program_management/business/excel.py
	modified:   program_management/business/group_element_years/group_element_year_tree.py
	modified:   program_management/tests/business/group_element_years/test_build_tree.py
