+++ base/tests/models/test_group_element_year.py
@@ -28,15 +28,11 @@ import random
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 
-from base.models import group_element_year
-from base.models.education_group_year import EducationGroupYear
-from base.models.enums import education_group_types
 from base.models.enums.education_group_types import GroupType, MiniTrainingType
 from base.models.enums.link_type import LinkTypes
 from base.models.group_element_year import GroupElementYear
 from base.tests.factories.academic_year import AcademicYearFactory
-from base.tests.factories.education_group_year import EducationGroupYearFactory, GroupFactory, MiniTrainingFactory, \
-    TrainingFactory
+from base.tests.factories.education_group_year import EducationGroupYearFactory, GroupFactory, MiniTrainingFactory
 from base.tests.factories.group_element_year import GroupElementYearFactory
 from base.tests.factories.learning_unit_year import LearningUnitYearFactory
 
@@ -121,67 +117,6 @@ class TestSaveGroupElementYear(TestCase):
                 child_branch=egy,
                 child_leaf=luy,
             )
-
-
-class TestFetchGroupElementsBehindHierarchy(TestCase):
-    """Unit tests on fetch_all_group_elements_behind_hierarchy()"""
-    @classmethod
-    def setUpTestData(cls):
-        cls.academic_year = AcademicYearFactory()
-        cls.root = TrainingFactory(
-            acronym='DROI2M',
-            education_group_type__name=education_group_types.TrainingType.PGRM_MASTER_120,
-            academic_year=cls.academic_year
-        )
-
-        finality_list = GroupFactory(
-            acronym='LIST FINALITIES',
-            education_group_type__name=education_group_types.GroupType.FINALITY_120_LIST_CHOICE,
-            academic_year=cls.academic_year
-        )
-
-        formation_master_md = TrainingFactory(
-            acronym='DROI2MD',
-            education_group_type__name=education_group_types.TrainingType.MASTER_MD_120,
-            academic_year=cls.academic_year
-        )
-
-        common_core = GroupFactory(
-            acronym='TC DROI2MD',
-            education_group_type__name=education_group_types.GroupType.COMMON_CORE,
-            academic_year=cls.academic_year
-        )
-
-        cls.link_1 = GroupElementYearFactory(parent=cls.root, child_branch=finality_list, child_leaf=None)
-        cls.link_1_bis = GroupElementYearFactory(parent=cls.root,
-                                                 child_branch=EducationGroupYearFactory(
-                                                     academic_year=cls.academic_year),
-                                                 child_leaf=None)
-        cls.link_2 = GroupElementYearFactory(parent=finality_list, child_branch=formation_master_md, child_leaf=None)
-        cls.link_3 = GroupElementYearFactory(parent=formation_master_md, child_branch=common_core, child_leaf=None)
-        cls.link_4 = GroupElementYearFactory(parent=common_core,
-                                             child_leaf=LearningUnitYearFactory(),
-                                             child_branch=None)
-
-    def test_with_one_root_id(self):
-        queryset = GroupElementYear.objects.all().select_related(
-            'child_branch__academic_year',
-            'child_leaf__academic_year',
-            # [...] other fetch
-        )
-        result = group_element_year.fetch_all_group_elements_in_tree(self.root, queryset)
-        expected_result = {
-            self.link_1.parent_id: [self.link_1, self.link_1_bis],
-            self.link_2.parent_id: [self.link_2],
-            self.link_3.parent_id: [self.link_3],
-            self.link_4.parent_id: [self.link_4],
-        }
-        self.assertDictEqual(result, expected_result)
-
-    def test_when_queryset_is_not_from_group_element_year_model(self):
-        wrong_queryset_model = EducationGroupYear.objects.all()
-        with self.assertRaises(AttributeError):
-            group_element_year.fetch_all_group_elements_in_tree(self.root, wrong_queryset_model)
 
 
 class TestValidationOnEducationGroupYearBlockField(TestCase):
